Include config-overrides.js in linter configuration
Some changes were required to the ESLint configuration so that it does
not have any issues with the CommonJS "require" statements used in
config-overrides.js.

To action these changes, all of the custom react-specific and import
rules were moved into a new .eslintrc inside src, leaving only the
react-app extends in the base .eslintrc.